Fix assignment operator bug
OrderedMap had a bug in it;s assignment operator which would
cause items deleted in the copy would delete corresponding
entries in the original ordered map.

This bug was caused due to OrderedMap storing iterators for
the linked list maintaining the insert order. Due to Qt
containers being implicitly shared, ot's not possible to
create copies on a container with active iterators. The
reason is that the iterator's maintain pointers to internal
state of the container and are hence any operation using them
will alter the state of the original container (even after
the container detaches, the iterator's still point to the
original container's state).

Fix (for now) is that when making a copy, invalidate old
iterators and regenrate new ones. This means copying
OrderedMap will take linear time.

Updated assignment operator unit test to have more coverage
and also be more correct.

TODO:
- Add copy constructor
- Enable implicit sharing