Fix obsoleted QLinkedList<T> and QList::fromStdList<T>
This is to silence warnings and suggestions to make these changes from
Qt 5.15 code - indeed the obsoleted things may not be present in Qt 6.x
at all.

I have also added licence details to the particular file I have edited and
inserted the copyright of those I have tracked as having made contributions
to it in the past.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>